chore: derived `Clone` (or also `Copy`) for most structs in the codebase. Changed type of sample rate from `i32` to `u32`.

+++ rustysynth_test/src/sample_util.rs
@@ -8,7 +8,7 @@ pub fn check(sample: &SampleHeader, values: &[i32; 7]) {
     assert_eq!(sample.get_end(), values[1]);
     assert_eq!(sample.get_start_loop(), values[2]);
     assert_eq!(sample.get_end_loop(), values[3]);
-    assert_eq!(sample.get_sample_rate(), values[4]);
+    assert_eq!(sample.get_sample_rate(), values[4] as u32);
     assert_eq!(sample.get_original_pitch(), values[5]);
     assert_eq!(sample.get_pitch_correction(), values[6]);
 }
